[fix] handle token refresh after logout
After a logout, making a token refresh attempt without this change means the execution reach the auth service. That one will attempt to verify whether the RT from the headers and the RT hash in the DB match. But the RT hash from the DB was deleted on logout => 500 error